+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Leaflet.Storage changelog
 
 ## dev
-- help tooltip now follows the mouse position
+- add tooltip when drawing
 
 ## 0.7.1
 - upgrade Leaflet.Editable to 0.2.0